* fixed all compilation issues of printkit-stuff with gcc4 on Linux * removed leagcy _sstream-header, we now use sstream, as both compilers provide it
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@19109 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/print/drivers/pdf/source/PageSetupWindow.cpp
@@ -164,7 +164,7 @@ PageSetupWindow::PageSetupWindow(BMessage *msg, const char *printerName)
 	margin.right = page.right - margin.right;
 	margin.bottom = page.bottom - margin.bottom;
 
-	fMarginView = new MarginView(BRect(20,20,200,160), width, height,
+	fMarginView = new MarginView(BRect(20,20,200,160), (int32)width, (int32)height,
 			margin, units);
 	panel->AddChild(fMarginView);
